+++ app/composables/ving/useVingRecord.mjs
@@ -1,7 +1,6 @@
 import { defineStore } from 'pinia';
 import { defu } from "defu";
 import { v4 } from 'uuid';
-import { ouch } from '#ving/utils/ouch.mjs';
 import { isObject, isUndefined } from '#ving/utils/identify.mjs';
 
 /**
@@ -28,12 +27,13 @@ import { isObject, isUndefined } from '#ving/utils/identify.mjs';
  * @param {object} behavior.links - Any default link objects. Defaults to `{}`. Usually no reason to ever specify this.
  * @param {object} behavior.related - Any default related objects. Defaults to `{}`. Usually no reason to ever specify this.
  * @param {object[]} behavior.warnings - An array of warning objects. Defaults to `[]`. Usually no reason to ever specify this.
+ * @param {string} behavior.ego An optional string that will be prepended to the id of fetched records in Pinia so that they can be distinguished from other instances of the same record. Useful if you're loading multiple instances of the same object on the same page.
  * @returns {VingRecordStore} - A Pinia store.
  * @example
  * const user = useVingRecord({
  *  id : 'xxx',
- *  fetchApi: '/api/v1/user/xxx',
- *  createApi: '/api/v1/user',
+ *  fetchApi: '/api/v1/users/xxx',
+ *  createApi: '/api/v1/users',
  *  query : { includeMeta : true },
  * });
  * await user.fetch();
@@ -42,8 +42,7 @@ import { isObject, isUndefined } from '#ving/utils/identify.mjs';
 
 export default (behavior) => {
     const notify = useNotify();
-
-    const generate = defineStore(behavior.id || v4(), {
+    const generate = defineStore(behavior.id ? (behavior.ego ? behavior.ego + '-' + behavior.id : behavior.id) : v4(), {
         state: () => ({
             props: behavior.props || {},
             meta: behavior.meta || {},
@@ -240,7 +239,7 @@ export default (behavior) => {
                     return this.links.base?.href;
                 }
                 notify.error('No createApi');
-                throw ouch(401, 'No createApi');
+                throw createError({ statusCode: 401, message: 'No createApi' });
             },
 
             /**
@@ -258,7 +257,7 @@ export default (behavior) => {
                     return this.links.self?.href;
                 }
                 notify.error('No fetchApi');
-                throw ouch(401, 'No fetchApi');
+                throw createError({ statusCode: 401, message: 'No fetchApi' });
             },
 
             /**
@@ -273,7 +272,7 @@ export default (behavior) => {
                     return this.links.self?.href;
                 }
                 notify.error('No links.self');
-                throw ouch(400, 'No links.self');
+                throw createError({ statusCode: 400, message: 'No links.self' });
             },
 
             /**